Improve util packaging
