+++ nat/factory_windows.go
@@ -18,7 +18,7 @@
 package nat
 
 import (
-	"github.com/mysteriumnetwork/node/utils"
+	"github.com/mysteriumnetwork/node/utils/cmdutil"
 )
 
 // NewService returns Windows OS specific NAT service based on Internet Connection Sharing (ICS).
@@ -26,6 +26,6 @@ func NewService() NATService {
 	return &serviceICS{
 		ifaces:          make(map[string]RuleForwarding),
 		setICSAddresses: setICSAddresses,
-		powerShell:      utils.PowerShell,
+		powerShell:      cmdutil.PowerShell,
 	}
 }
